Fix syntax to satisfy Sequras exigences

+++ db/migrate/20220726151753_create_disbursements.rb
@@ -3,7 +3,7 @@ class CreateDisbursements < ActiveRecord::Migration[7.0]
     create_table :disbursements do |t|
       t.datetime :date, null: false
       t.integer :merchant_id, null: false
-      t.float :amount
+      t.decimal :amount
 
       t.timestamps
     end
